Improves styling and testing
Refactors styling with a new `sjOutlinedCard` style.

Updates the test command name for better clarity.

Also adds standalone component support to the demo app components.

Improves the `SjDirective` to handle pseudo-selectors and nested objects and simplifies style merging using deep merge for better theme consistency.

Adds a test case to ensure the CSS cache is cleared and the theme version is bumped when the theme changes.