Introduces SjHostComponent for enhanced styling
Adds the `SjHostComponent` to provide a more flexible and encapsulated way to apply Super-JSS styles.

This component acts as a host for content, allowing styles to be applied to the parent element or a specific child element. This approach improves component composition and avoids potential conflicts with other directives.

Refactors the demo app to utilize `SjHostComponent` for header, sidenav, and main application layout, streamlining styling and improving code readability.