Refactors demo cards for improved structure
Improves the structure and styling of the demo cards component.

This change replaces the previous flexbox-based layout with a grid layout for better responsiveness and organization. It also leverages the `sj-card` component and `SectionContainerComponent` for a cleaner and more maintainable structure. The `info` variant of the card is updated to use the 'info' color from the palette and the interactive button style is updated to use the blueprint.