Adds `usePaint` prop to button component
Adds the `usePaint` property to the button component.

This new property allows for specifying a background and text color based on a theme token (e.g., "primary.main" for background and "primary.contrast" for text).
It gives SuperJSS button more flexibility in styling and improves the user experience.

Also updates documentation and removes unused card and stack stories.
Increments the package version.